+++ process/sync/baseSync_test.go
@@ -34,87 +34,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-type requestHandlerWithSetEpochStub struct {
-	testscommon.RequestHandlerStub
-	SetEpochCalled func(epoch uint32)
-}
-
-func (rhs *requestHandlerWithSetEpochStub) SetEpoch(epoch uint32) {
-	if rhs.SetEpochCalled != nil {
-		rhs.SetEpochCalled(epoch)
-		return
-	}
-
-	rhs.RequestHandlerStub.SetEpoch(epoch)
-}
-
-type blockBootstrapperStub struct {
-	getCurrHeaderCalled       func() (data.HeaderHandler, error)
-	getPrevHeaderCalled       func(data.HeaderHandler, storage.Storer) (data.HeaderHandler, error)
-	getBlockBodyCalled        func(data.HeaderHandler) (data.BodyHandler, error)
-	isForkTriggeredByMetaFunc func() bool
-	requestProofByNonceCalled func(nonce uint64)
-}
-
-func (bbs *blockBootstrapperStub) getCurrHeader() (data.HeaderHandler, error) {
-	if bbs.getCurrHeaderCalled != nil {
-		return bbs.getCurrHeaderCalled()
-	}
-	return nil, nil
-}
-
-func (bbs *blockBootstrapperStub) getPrevHeader(header data.HeaderHandler, storer storage.Storer) (data.HeaderHandler, error) {
-	if bbs.getPrevHeaderCalled != nil {
-		return bbs.getPrevHeaderCalled(header, storer)
-	}
-	return nil, nil
-}
-
-func (bbs *blockBootstrapperStub) getBlockBody(header data.HeaderHandler) (data.BodyHandler, error) {
-	if bbs.getBlockBodyCalled != nil {
-		return bbs.getBlockBodyCalled(header)
-	}
-	return nil, nil
-}
-
-func (bbs *blockBootstrapperStub) getHeaderWithHashRequestingIfMissing([]byte) (data.HeaderHandler, error) {
-	return nil, nil
-}
-
-func (bbs *blockBootstrapperStub) getHeaderWithNonceRequestingIfMissing(uint64) (data.HeaderHandler, []byte, error) {
-	return nil, nil, nil
-}
-
-func (bbs *blockBootstrapperStub) getBlockBodyRequestingIfMissing(data.HeaderHandler) (data.BodyHandler, error) {
-	return nil, nil
-}
-
-func (bbs *blockBootstrapperStub) isForkTriggeredByMeta() bool {
-	if bbs.isForkTriggeredByMetaFunc != nil {
-		return bbs.isForkTriggeredByMetaFunc()
-	}
-	return false
-}
-
-func (bbs *blockBootstrapperStub) requestHeaderByNonce(uint64) {
-}
-
-func (bbs *blockBootstrapperStub) requestProofByNonce(nonce uint64) {
-	if bbs.requestProofByNonceCalled != nil {
-		bbs.requestProofByNonceCalled(nonce)
-	}
-}
-
-func getMockChainHandler() data.ChainHandler {
-	return &testscommon.ChainHandlerStub{
-		GetGenesisHeaderCalled: func() data.HeaderHandler {
-			return &block.Header{
-				Epoch: 0,
-			}
-		},
-	}
-}
-
 func TestBaseBootstrap_SyncBlocksShouldNotCallSyncIfNotConnectedToTheNetwork(t *testing.T) {
 	t.Parallel()
 
@@ -933,23 +852,23 @@ func TestBaseBootstrap_SaveProposedTxsToPool(t *testing.T) {
 	header := &block.HeaderV3{}
 	body := &block.Body{
 		MiniBlocks: []*block.MiniBlock{
-			&block.MiniBlock{
+			{
 				TxHashes: [][]byte{[]byte("txHash1")},
 				Type:     block.TxBlock,
 			},
-			&block.MiniBlock{
+			{
 				TxHashes: [][]byte{[]byte("txHash2")},
 				Type:     block.InvalidBlock,
 			},
-			&block.MiniBlock{
+			{
 				TxHashes: [][]byte{[]byte("txHash3")},
 				Type:     block.SmartContractResultBlock,
 			},
-			&block.MiniBlock{
+			{
 				TxHashes: [][]byte{[]byte("txHash4")},
 				Type:     block.RewardsBlock,
 			},
-			&block.MiniBlock{
+			{
 				TxHashes: [][]byte{[]byte("txHash5")},
 				Type:     block.PeerBlock,
 			},
